refactor: move infrequently used modal.Function into utils module
